Remove authorisation for old-style image uploads
Previously all github image uploads were public, even on private
repos. The host of the uploads was user-images.githubusercontent.com
and accessible without any authorisation.

This was changed in May 2023 [0] whereby image uploads are tied
to the repository and have the normal owner/repo structure as the url
path.

Trying to access the public images with an authorization header set
causes an error, removing the header works fine.

[0] https://github.blog/changelog/2023-05-09-more-secure-private-attachments/